refactor git page scripts
